+++ core/imds/src/main/java/software/amazon/awssdk/imds/MetadataResponse.java
@@ -28,7 +28,7 @@ import software.amazon.awssdk.utils.ToString;
 import software.amazon.awssdk.utils.Validate;
 
 /**
- * The class is used for Response Handling and Parsing the metadata fetched by the get call in the {@link Ec2MetadataClient}
+ * The class is used for response handling and parsing the metadata fetched by the get call in the {@link Ec2MetadataClient}
  * interface.
  * The class provides convenience methods to the users to parse the metadata as a String, List and Document.
  */
