+++ core/imds/src/main/java/software/amazon/awssdk/imds/internal/Ec2MetadataEndpointProvider.java
@@ -57,7 +57,7 @@ public final class Ec2MetadataEndpointProvider {
 
     /**
      * Resolve the endpoint to be used for the {@link DefaultEc2MetadataClient} client. Users may manually provide an endpoint
-     * through the {@code AWS_EC2_METADATA_SERVICE_ENDPOINT} environment variable or th {@code ec2_metadata_service_endpoint}
+     * through the {@code AWS_EC2_METADATA_SERVICE_ENDPOINT} environment variable or the {@code ec2_metadata_service_endpoint}
      * key in
      * their aws config file.
      * If an endpoint is specified is this manner, use it. If no value are provide, the defaults to:
